+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - `remove double error logging if fragments dir is not found`
 - `rename fragments.list_fragments to fragments.fragments_list`
 - `rename fragments.list_loaded to fragments.loaded_list`
+- `rename schemas.list_schemas to schemas.schemas_list`
 
 ## 0.0.1